+++ src/animation/use-animation.ts
@@ -1,6 +1,5 @@
 import { AnimationControls } from "./AnimationControls"
 import { useMemo, useEffect } from "react"
-import { Transition, Variants } from "../types"
 
 /**
  * The `useAnimation` hook returns `AnimationControls`, which can be used to manually start, stop and sequence animations on one or more Frames.
@@ -13,42 +12,29 @@ import { Transition, Variants } from "../types"
  * import { Frame, useAnimation } from "framer"
  *
  * export function MyComponent(props) {
- *    const animation = useAnimation()
+ *    const controls = useAnimation()
  *
- *    animation.start({
+ *    controls.start({
  *        x: 100,
  *        transition: { duration: 0.5 },
  *    })
  *
- *    return <Frame animate={animation} />
+ *    return <Frame animate={controls} />
  * }
  *
  * ```
  *
- * @param variants - An optional named map of variants.
- * @param defaultTransition - An optional default transition to use when a variant doesn’t have an explicit `transition` property set.
  * @returns Animation controller with `start` and `stop` methods
  *
  * @public
  */
-export function useAnimation(
-    variants?: Variants,
-    defaultTransition?: Transition
-) {
-    const animationManager = useMemo(() => new AnimationControls(), [])
-
-    if (variants) {
-        animationManager.setVariants(variants)
-    }
-
-    if (defaultTransition) {
-        animationManager.setDefaultTransition(defaultTransition)
-    }
+export function useAnimation() {
+    const animationControls = useMemo(() => new AnimationControls(), [])
 
     useEffect(() => {
-        animationManager.mount()
-        return () => animationManager.unmount()
+        animationControls.mount()
+        return () => animationControls.unmount()
     }, [])
 
-    return animationManager
+    return animationControls
 }